chore: cleanup for consistency
