--- config-tests.toml
@@ -1,52 +1,0 @@
-# ===========================
-# ZeroXBridge Sequencer Config
-# ===========================
-
-[contract]
-name = "zeroXBridge"
-
-[contracts]
-l1_contract_address = "0x0000000000000000000000000000000000000000"  # Replace with actual L1 contract
-l2_contract_address = "0x0000000000000000000000000000000000000000"  # Replace with actual L2 contract
-
-[server]
-host = "127.0.0.1"
-server_url = "http://127.0.0.1:4000"
-
-[database]
-max_connections = 10
-
-[ethereum]
-chain_id = 1
-confirmations = 3
-
-[starknet]
-chain_id = "0x534e5f4d41494e"  # SN_MAIN
-
-[relayer]
-max_retries = 5
-retry_delay_seconds = 10
-gas_limit = 500000
-
-[queue]
-process_interval_sec = 5
-wait_time_seconds = 5
-max_retries = 3
-initial_retry_delay_sec = 10
-retry_delay_seconds = 15
-merkle_update_confirmations = 5
-
-[merkle]
-tree_depth = 32
-cache_size = 1000
-
-[logging]
-level = "info"              # Options: debug, info, warn, error
-file = "logs/sequencer.log"
-
-[oracle]
-tolerance_percent = 0.01    # 1%
-polling_interval_seconds = 60
-
-[herodotus]
-herodotus_endpoint = "https://herodotus.example.com/api"
