+++ tests/l2_event_watcher.rs
@@ -1,15 +1,13 @@
 use anyhow::Result;
 use mockall::predicate::*;
 use mockall::*;
-use sqlx::postgres::PgPoolOptions;
-use sqlx::PgPool;
 use starknet::core::types::{EmittedEvent, EventFilter, EventsPage, Felt};
-use zeroxbridge_sequencer::config::{
-    AppConfig, ContractConfig, Contracts, DatabaseConfig, EthereumConfig, LoggingConfig,
-    MerkleConfig, OracleConfig, QueueConfig, RelayerConfig, ServerConfig, StarknetConfig,
-};
+
 use zeroxbridge_sequencer::events::fetch_l2_burn_events;
 use zeroxbridge_sequencer::events::l2_event_watcher::TestProvider;
+
+#[path = "utils.rs"]
+mod utils;
 
 // Mock the Starknet Provider
 mock! {
@@ -45,74 +43,8 @@ impl TestProvider for MockStarknetProvider {
 // Test module to group all L2 event watcher tests
 #[cfg(test)]
 mod tests {
-    use zeroxbridge_sequencer::config::HerodotusConfig;
-
     use super::*;
-
-    // Helper function to create test database pool
-    async fn setup_test_db() -> PgPool {
-        let database_url = std::env::var("DATABASE_URL")
-            .unwrap_or_else(|_| "postgres://postgres:postgres@localhost:5435/zeroxdb".to_string());
-
-        PgPoolOptions::new()
-            .max_connections(1)
-            .connect(&database_url)
-            .await
-            .expect("Failed to connect to test database")
-    }
-
-    // Helper function to create test config
-    pub fn create_test_config() -> AppConfig {
-        AppConfig {
-            contract: ContractConfig {
-                name: String::new(),
-            },
-            contracts: Contracts {
-                l1_contract_address: "0x123".to_string(),
-                l2_contract_address: "0x456".to_string(),
-            },
-            server: ServerConfig {
-                host: "127.0.0.1".to_string(),
-                server_url: "http://localhost:8080".to_string(),
-            },
-            database: DatabaseConfig { max_connections: 5 },
-            ethereum: EthereumConfig {
-                chain_id: 11155111, // Sepolia testnet
-                confirmations: 1,
-            },
-            starknet: StarknetConfig {
-                chain_id: "SN_SEPOLIA".to_string(),
-            },
-            relayer: RelayerConfig {
-                max_retries: 3,
-                retry_delay_seconds: 60,
-                gas_limit: 300000,
-            },
-            queue: QueueConfig {
-                process_interval_sec: 60,
-                wait_time_seconds: 30,
-                max_retries: 3,
-                initial_retry_delay_sec: 60,
-                retry_delay_seconds: 60,
-                merkle_update_confirmations: 1,
-            },
-            merkle: MerkleConfig {
-                tree_depth: 32,
-                cache_size: 1000,
-            },
-            logging: LoggingConfig {
-                level: "info".to_string(),
-                file: "logs/zeroxbridge.log".to_string(),
-            },
-            oracle: OracleConfig {
-                tolerance_percent: Some(0.01), // 1% tolerance
-                polling_interval_seconds: 60,
-            },
-            herodotus: HerodotusConfig {
-                herodotus_endpoint: "23478234".to_string(),
-            },
-        }
-    }
+    use utils::create_test_app;
 
     // Helper function to create a test event
     fn create_test_event(
@@ -143,8 +75,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_commitment_hash_decoding() -> Result<()> {
-        let pool = setup_test_db().await;
-        let config = create_test_config();
+        let app = create_test_app().await;
         let mut mock_provider = MockStarknetProvider::new();
 
         // Mock block number response
@@ -164,7 +95,7 @@ mod tests {
             })
         });
 
-        let result = fetch_l2_burn_events(&config, &pool, 90, &mock_provider).await?;
+        let result = fetch_l2_burn_events(&app.config, &app.db, 90, &mock_provider).await?;
 
         assert_eq!(result.len(), 2);
         assert_eq!(result[0].commitment_hash, "0xabc");
@@ -179,8 +110,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_block_index_tracking() -> Result<()> {
-        let pool = setup_test_db().await;
-        let config = create_test_config();
+        let app = create_test_app().await;
         let mut mock_provider = MockStarknetProvider::new();
 
         mock_provider.expect_block_number().returning(|| Ok(100));
@@ -205,14 +135,14 @@ mod tests {
             });
 
         // First call: should process blocks 90-95
-        let result = fetch_l2_burn_events(&config, &pool, 90, &mock_provider).await?;
+        let result = fetch_l2_burn_events(&app.config, &app.db, 90, &mock_provider).await?;
         assert_eq!(result.len(), 1);
 
         // Verify block tracker was updated
         let last_block = sqlx::query!(
             "SELECT last_block FROM block_trackers WHERE key = 'l2_burn_events_last_block'"
         )
-        .fetch_one(&pool)
+        .fetch_one(&app.db)
         .await?;
 
         assert_eq!(result[0].block_number, 95);
@@ -223,8 +153,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_large_amount_handling() -> Result<()> {
-        let pool = setup_test_db().await;
-        let config = create_test_config();
+        let app = create_test_app().await;
         let mut mock_provider = MockStarknetProvider::new();
 
         mock_provider.expect_block_number().returning(|| Ok(100));
@@ -245,7 +174,7 @@ mod tests {
             })
         });
 
-        let result = fetch_l2_burn_events(&config, &pool, 92, &mock_provider).await?;
+        let result = fetch_l2_burn_events(&app.config, &app.db, 92, &mock_provider).await?;
 
         assert_eq!(result.len(), 1);
         assert_eq!(result[0].commitment_hash, "0xabc");
